check user enable status

+++ handlers/auth.go
@@ -106,6 +106,13 @@ func AuthBasic(c *gin.Context) {
 		}
 		db.Create(&existingUser)
 	}
+	if !existingUser.Enable {
+		c.JSON(401, ResponseHTTP{
+			Success: false,
+			Message: "User is disabled",
+		})
+		return
+	}
 
 	if existingUser.GiteaToken == "" {
 		client.DeleteAccessToken("OJ-API")
@@ -240,10 +247,17 @@ func RefreshToken(c *gin.Context) {
 
 	// Check if user exists and refresh token matches
 	var user models.User
-	if err := db.Where("id = ? AND refresh_token = ?", claims.UserID, refreshToken).First(&user).Error; err != nil {
+	if err := db.Where("id = ? AND refresh_token = ?", claims.UserID, refreshToken, true).First(&user).Error; err != nil {
 		c.JSON(401, ResponseHTTP{
 			Success: false,
 			Message: "Invalid refresh token",
+		})
+		return
+	}
+	if !user.Enable {
+		c.JSON(401, ResponseHTTP{
+			Success: false,
+			Message: "User is disabled",
 		})
 		return
 	}
